refactor(types): extract inventory receipt DTOs
